Test setup for pages. Large test for campaigns
Campaign functionality is a very rigorous change, and would have
benefited from an existing suite of exhaustive tests, but sadly, these
do not exist. As such, a rather complicated scenario that touches all
campaign-related pages was devised.

The scenario, is, as of yet, unfinished, as Compendium, Playlist and
Token functionality are not yet part of the test, nor are the pages for
adding tokens through the map interface, as well as the exploration and
combat interface.